Improve fountain import export. Issue #2
Resolves https://github.com/juliaclement/trelby/issues/2

On export, send lines from title pages to the .fountain file

On import, read entries in the title page section and add them to
page 0 of the title pages.

Title, Author and Title are matched to the default strings trelby
creates. Other lines in format kw: data are added as custom entries